requirements.txt
openai
easyocr
torch
playwright
twocaptcha_extension_python
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib